Self-Guided Learning/docker/day_02/examples/simple_project/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "./src" }]

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.29.0"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
